+++ src/assets/about/wildhacks-opening.png
