+++ src/assets/about/wildhacks-team-pose.jpg
