+++ src/assets/about/wildhacks-teamwork.jpg
