+++ src/assets/sponsors/northwestern-cs.png
